requirements.txt
numpy
jupyter
jupytext
pint
tqdm
pandas
python-dotenv
dspy
datasets
filelock
langchain
ortools
matplotlib
seaborn
pydantic
quantulum3
dateparser
parsedatetime
stanza
spacy
timexy
litellm
strsimpy
bert_score